some basic error handling

+++ src/main/java/pl/pw/edu/STOD_Warehouse/services/ClubEventService.java
@@ -11,21 +11,25 @@ public class ClubEventService {
     @Autowired
     private ClubEventRepository clubEventRepository;
 
-    public long addClubEvent(String name, String headquarterId){
-        return clubEventRepository.save(new ClubEvent(0L,name,headquarterId)).getId();
+    public long addClubEvent(String name, String headquarterId) {
+        return clubEventRepository.save(new ClubEvent(0L, name, headquarterId)).getId();
     }
 
-    public long deleteClubEvent(String name){
-        ClubEvent toDelete = clubEventRepository.findByName(name);
-        clubEventRepository.delete(toDelete);
-        return 0L;
+    public long deleteClubEvent(Long id) {
+        if (clubEventRepository.findById(id).isPresent()) {
+            ClubEvent toDelete = clubEventRepository.findById(id).get();
+            clubEventRepository.delete(toDelete);
+            return 0L;
+        } else {
+            return 1L;
+        }
     }
 
-    public ClubEvent getOne(Long id){
+    public ClubEvent getOne(Long id) {
         return clubEventRepository.getOne(id);
     }
 
-    public List<ClubEvent> getAllClubEvents(){
+    public List<ClubEvent> getAllClubEvents() {
         return clubEventRepository.findAll();
     }
